refactor(argocd): remove headlamp serviceaccount
This commit removes the headlamp service account from the
Argocd application. This is because the service account is
no longer needed and simplifies the application.

refactor(argocd): rename keycloak admin environment variables

This commit renames the Keycloak admin environment variables
to use the standard Keycloak naming convention. This improves
consistency and readability.